Removed redeclaration of struct and moved it to a common header
