+++ tests/greybus/integration/audio/CMakeLists.txt
@@ -5,6 +5,7 @@ set(DTS_ROOT ${CMAKE_CURRENT_SOURCE_DIR})
 project(fake_audio_test)
 
 get_filename_component(GB_ROOT ${CMAKE_CURRENT_SOURCE_DIR}/../../../.. ABSOLUTE)
+target_include_directories(app PRIVATE ${GB_ROOT}/drivers/audio)
 
 target_sources(app PRIVATE
   src/main.c
